refactor: rewrite a centralized experiment pipeline and pipe out results into latex compatible element
