+++ .gitignore
@@ -89,3 +89,6 @@ __pycache__/
 .venv
 env/
 venv/
+
+**/.DS_Store
+
