--- non-temporal/smoteenn_outputs/baseline_vs_smoteenn_comparison.csv
@@ -1,6 +1,0 @@
-Model,Baseline ROC-AUC,SMOTEENN ROC-AUC,ΔAUC,Baseline Recall (Fail),SMOTEENN Recall (Fail),ΔRecall (Fail)
-Logistic Regression,0.6806740614334471,0.6815499963691816,0.000875934935734457,0.6329787234042553,0.7712765957446809,0.13829787234042556
-Random Forest,0.6831248638443106,0.7153574540701475,0.03223259022583691,0.324468085106383,0.5797872340425532,0.2553191489361702
-Gradient Boosting,0.6948932539394379,0.7121328334906687,0.017239579551230788,0.0851063829787234,0.44680851063829785,0.36170212765957444
-XGBoost,0.6778896775833273,0.7095141057294314,0.03162442814610411,0.5638297872340425,0.4787234042553192,-0.08510638297872336
-LightGBM,0.678270913513906,0.7118695991576501,0.03359868564374413,0.5425531914893617,0.48404255319148937,-0.058510638297872286
