--- non-temporal/smoteenn_outputs/cv_smoteenn_results.csv
@@ -1,6 +1,0 @@
-Model,CV Accuracy,CV F1,CV ROC-AUC,CV Recall,CV Precision
-Logistic Regression +SMOTEENN,0.5438 ± 0.0095,0.6574 ± 0.0089,0.6947 ± 0.0097,0.5078 ± 0.0094,0.9319 ± 0.0044
-Random Forest +SMOTEENN,0.7088 ± 0.0139,0.8123 ± 0.0109,0.7137 ± 0.0110,0.7315 ± 0.0171,0.9135 ± 0.0014
-Gradient Boosting +SMOTEENN,0.7547 ± 0.0116,0.8492 ± 0.0083,0.7096 ± 0.0107,0.8020 ± 0.0145,0.9026 ± 0.0022
-XGBoost +SMOTEENN,0.7484 ± 0.0144,0.8446 ± 0.0108,0.7110 ± 0.0105,0.7941 ± 0.0201,0.9022 ± 0.0033
-LightGBM +SMOTEENN,0.7540 ± 0.0116,0.8493 ± 0.0089,0.7058 ± 0.0075,0.8049 ± 0.0183,0.8993 ± 0.0030
