--- non-temporal/smoteenn_outputs/smoteenn_model_results.csv
@@ -1,6 +1,0 @@
-Model,Accuracy,ROC-AUC,F1 (Failure),Recall (Failure),Precision (Failure),F1 (Success),Recall (Success),Train Time (s),Iterations / Trees
-Random Forest + SMOTEENN,0.7073529411764706,0.7153574540701475,0.3538961038961039,0.5797872340425532,0.2546728971962617,0.8108365019011406,0.7278156996587031,0.95,300
-Gradient Boosting + SMOTEENN,0.7448529411764706,0.7121328334906687,0.3262135922330097,0.44680851063829785,0.25688073394495414,0.8426303854875283,0.7926621160409556,3.51,300
-LightGBM + SMOTEENN,0.7477941176470588,0.7118695991576501,0.3466666666666667,0.48404255319148937,0.27002967359050445,0.8437357630979498,0.7901023890784983,1.43,300
-XGBoost + SMOTEENN,0.7448529411764706,0.7095141057294314,0.3415559772296015,0.4787234042553192,0.26548672566371684,0.8417692658458732,0.787542662116041,1.41,300
-Logistic Regression + SMOTEENN,0.5095588235294117,0.6815499963691816,0.30303030303030304,0.7712765957446809,0.18855656697009102,0.6216676120249575,0.46757679180887374,0.53,11
