+++ paper/apa/tables/top_models_comparison.tex
@@ -1,0 +1,23 @@
+\begin{table}[htbp]
+\caption{Top 10 Performing Model Configurations Across All Experiments}
+\label{tab:top_models}
+\begin{tabular}{lcccc}
+\toprule
+
+Sampler & Model & ROC-AUC & Recall (Fail) & F1 (Fail) \\
+\midrule
+
+SMOTE - Tomek & Random Forest & 0.7195 & 0.4043 & 0.3510 \\
+SMOTE & Random Forest & 0.7137 & 0.3883 & 0.3364 \\
+SMOTE & LightGBM & 0.7136 & 0.2021 & 0.2542 \\
+SMOTE - ENN & LightGBM & 0.7136 & 0.4202 & 0.3473 \\
+SMOTE - ENN & Gradient Boosting & 0.7125 & 0.4362 & 0.3241 \\
+SMOTE - ENN & Random Forest & 0.7109 & 0.5160 & 0.3328 \\
+SMOTE - Tomek & XGBoost & 0.7104 & 0.7553 & 0.3369 \\
+SMOTE - Tomek & LightGBM & 0.7093 & 0.2181 & 0.2680 \\
+SMOTE & XGBoost & 0.7075 & 0.7287 & 0.3235 \\
+SMOTE & Gradient Boosting & 0.7051 & 0.2128 & 0.2721 \\
+\bottomrule
+
+\end{tabular}
+\end{table}
